@@ -7,8 +7,8 @@ import android.content.res.Resources;
 import android.graphics.Point;
 import android.graphics.drawable.GradientDrawable;
 import android.os.Build;
-import android.support.design.widget.TabLayout;
-import android.support.v4.content.ContextCompat;
+import com.google.android.material.tabs.TabLayout;
+import androidx.core.content.ContextCompat;
 import android.view.Display;
 import android.view.View;
 import android.view.ViewGroup;
