@@ -3,18 +3,14 @@ package com.qartf.doseforreddit.view.activity;
 import android.content.SharedPreferences;
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.design.widget.TabLayout;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
 
+import com.google.android.material.navigation.NavigationView;
+import com.google.android.material.tabs.TabLayout;
 import com.qartf.doseforreddit.R;
 import com.qartf.doseforreddit.presenter.comment.CommentMVP;
 import com.qartf.doseforreddit.presenter.post.PostMVP;
@@ -29,6 +25,10 @@ import com.qartf.doseforreddit.view.dialog.LoginDialog;
 
 import javax.inject.Inject;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.widget.Toolbar;
+import androidx.drawerlayout.widget.DrawerLayout;
 import butterknife.BindString;
 import butterknife.BindView;
 
