@@ -1,7 +1,7 @@
 package com.qartf.doseforreddit.view.fragment;
 
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.TabLayout;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.tabs.TabLayout;
 import android.text.Editable;
 import android.text.InputFilter;
 import android.text.TextUtils;
